create-product

+++ src/products/entities/product.entity.ts
@@ -12,7 +12,7 @@ export class Product {
   })
   title: string;
 
-  @Column('numeric', {
+  @Column('float', {
     default: 0,
   })
   price: number;
@@ -41,6 +41,6 @@ export class Product {
   @Column('text')
   gender: string;
 
-  //   tags
+   //   tags
   // images
 }
